Environment utility (env) moved to seperate directory
